limit access to views to users in enclosures
refactor code out of views which aren't views
